+++ changelog.md
@@ -43,3 +43,4 @@
 ## 0.0.6
 
 - Added demos.
+- Added num_sites, theta_w to sim()